requirements.txt
pandas
requests
great_expectations
psycopg2-binary
streamlit
pandera